function/requirements.txt
google-cloud-storage==3.1.0
Flask==3.1.0
